requirements.txt
streamlit
mysql.connector
python-dotenv
openai
instructor
pydantic
st-paywall
